+++ docs/.vitepress/theme/index.ts
@@ -20,6 +20,7 @@ import DemoGkField from '../components/demos/form/field/DemoGkField.vue'
 import DemoGkInput from '../components/demos/form/input/DemoGkInput.vue'
 import DemoGkRadio from '../components/demos/form/radio/DemoGkRadio.vue'
 import DemoGkSelect from '../components/demos/form/select/DemoGkSelect.vue'
+import DemoGkSnackbar from '../components/demos/feedback/snackbar/DemoGkSnackbar.vue'
 import DemoGkSkeletonLoader from '../components/demos/feedback/skeleton-loader/DemoGkSkeletonLoader.vue'
 import DemoGkSpinner from '../components/demos/spinner/DemoGkSpinner.vue'
 import DemoGkStack from '../components/demos/stack/DemoGkStack.vue'
@@ -51,6 +52,7 @@ export default {
     app.component('DemoGkContainer', DemoGkContainer)
     app.component('DemoGkExpansionPanels', DemoGkExpansionPanels)
     app.component('DemoGkDivider', DemoGkDivider)
+    app.component('DemoGkSnackbar', DemoGkSnackbar)
     app.component('DemoGkSkeletonLoader', DemoGkSkeletonLoader)
     app.component('DemoGkSpinner', DemoGkSpinner)
   },